Made some crucial changes
