+++ cmake-build-release/.ninja_log
@@ -1,40 +1,35 @@
 # ninja log v5
+7964	11672	7211490380769982	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/input.c.obj	bb60a92d2be5f097
+3	375	7211975565141464	CMakeFiles/GithubGameoff.dir/main.cpp.obj	13b75ed8f7632aba
+376	489	7211975566235063	GithubGameoff.exe	e64f240aa221ff81
 2	2348	7211490287124081	CMakeFiles/GithubGameoff.dir/gameobject.cpp.obj	b45a971e92c1de4a
 62	2393	7211490287404164	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/rshapes.c.obj	906844b80b21a197
-2352	5400	7211490318246842	CMakeFiles/GithubGameoff.dir/main.cpp.obj	13b75ed8f7632aba
+18623	22456	7211490488666036	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/egl_context.c.obj	ace2082043e264cd
 39	5563	7211490319614067	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/rtext.c.obj	c74516fd6a980fab
-5401	7964	7211490343413769	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/utils.c.obj	f66d930bb8459e89
 5564	11653	7211490380553029	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/context.c.obj	92b3d019cc6ce2e
-7964	11672	7211490380769982	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/input.c.obj	bb60a92d2be5f097
+12700	15120	7211490414468700	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/platform.c.obj	4b734efaa35e6ee8
+5401	7964	7211490343413769	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/utils.c.obj	f66d930bb8459e89
 30	12699	7211490390665256	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/rcore.c.obj	898fd45dd02986e0
 11656	14907	7211490413193983	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/vulkan.c.obj	b77278ee0fc54af9
 11673	14939	7211490413407492	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/window.c.obj	fa255e611b73d8d4
-12700	15120	7211490414468700	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/platform.c.obj	4b734efaa35e6ee8
 14907	18513	7211490449253863	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/monitor.c.obj	c8024c39f2e89aad
 14940	18526	7211490449330109	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/init.c.obj	1bbb5f3a2352dde9
 15121	18622	7211490450399761	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/null_window.c.obj	6ea49867959309fb
+50	32487	7211490589237913	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/rtextures.c.obj	19bd02f25e5ddc25
 18514	21456	7211490478923417	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/null_init.c.obj	388044cd08a59973
 18527	21470	7211490478891265	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/null_joystick.c.obj	862837730f386a2a
-18623	22456	7211490488666036	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/egl_context.c.obj	ace2082043e264cd
+2394	50083	7211490765182310	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/raudio.c.obj	1fade9d5ca13793c
 21457	24291	7211490507252296	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/osmesa_context.c.obj	3f8cdeac0b17ef77
 21472	24321	7211490507497005	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/null_monitor.c.obj	14908198ecd4b593
+14	1620	7211977495941367	build.ninja	bf6299c1d472140c
 22456	27383	7211490538191137	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/win32_joystick.c.obj	485e845247083b82
 24292	27908	7211490543050152	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/win32_init.c.obj	68faa43dc6a982fb
 13	29854	7211490561242327	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/rmodels.c.obj	a828afcb2dc31dbc
 24321	29865	7211490561748917	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/win32_monitor.c.obj	8116f2b35bf89c95
+29875	32867	7211490593031276	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/wgl_context.c.obj	d2e9ef40308e58c4
 27384	29874	7211490562714308	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/win32_thread.c.obj	541334da1596eee7
 27909	31545	7211490578639435	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/win32_module.c.obj	fea853541e692bdf
 29854	32358	7211490587846510	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/win32_time.c.obj	204b7a47d65a17fc
-50	32487	7211490589237913	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/rtextures.c.obj	19bd02f25e5ddc25
-29875	32867	7211490593031276	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/wgl_context.c.obj	d2e9ef40308e58c4
 29865	33432	7211490598702079	_deps/raylib-build/raylib/external/glfw/src/CMakeFiles/glfw.dir/win32_window.c.obj	867f60e44e3f7a66
 33432	33603	7211490600407278	_deps/raylib-build/raylib/external/glfw/src/libglfw3.a	ee173a4d6a9b4784
-2394	50083	7211490765182310	_deps/raylib-build/raylib/CMakeFiles/raylib.dir/raudio.c.obj	1fade9d5ca13793c
 50087	50212	7211490766519967	_deps/raylib-build/raylib/libraylib.a	f1b314a44adff93a
-50212	50493	7211490769166911	GithubGameoff.exe	e64f240aa221ff81
-5	902	7211491310283618	CMakeFiles/GithubGameoff.dir/main.cpp.obj	13b75ed8f7632aba
-903	1147	7211491312661904	GithubGameoff.exe	e64f240aa221ff81
-10	770	7211492630580989	CMakeFiles/GithubGameoff.dir/main.cpp.obj	13b75ed8f7632aba
-771	1009	7211492632884095	GithubGameoff.exe	e64f240aa221ff81
-7	805	7211492854954811	CMakeFiles/GithubGameoff.dir/main.cpp.obj	13b75ed8f7632aba
-806	1048	7211492857306688	GithubGameoff.exe	e64f240aa221ff81
