+++ cmake-build-release/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Nov 09 07:03 India Standard Time
+Start testing: Nov 09 07:19 India Standard Time
 ----------------------------------------------------------
-End testing: Nov 09 07:03 India Standard Time
+End testing: Nov 09 07:19 India Standard Time
