+++ cmake-build-release/_deps/raylib_cpp-subbuild/build.ninja
@@ -173,14 +173,14 @@ build all: phony raylib_cpp-populate
 #############################################
 # Re-run CMake if any of its inputs changed.
 
-build build.ninja: RERUN_CMAKE | C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeDetermineSystem.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeGenericSystem.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeInitializeConfigs.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystem.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystemSpecificInformation.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystemSpecificInitialize.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\RepositoryInfo.txt.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\cfgcmd.txt.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\gitclone.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\gitupdate.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\mkdirs.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\Platform\Windows.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\Platform\WindowsPaths.cmake CMakeCache.txt CMakeFiles\3.26.4\CMakeSystem.cmake CMakeLists.txt raylib_cpp-populate-prefix\tmp\raylib_cpp-populate-mkdirs.cmake
+build build.ninja: RERUN_CMAKE | C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeGenericSystem.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeInitializeConfigs.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystemSpecificInformation.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystemSpecificInitialize.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\RepositoryInfo.txt.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\cfgcmd.txt.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\gitclone.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\gitupdate.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\mkdirs.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\Platform\Windows.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\Platform\WindowsPaths.cmake CMakeCache.txt CMakeFiles\3.26.4\CMakeSystem.cmake CMakeLists.txt raylib_cpp-populate-prefix\tmp\raylib_cpp-populate-mkdirs.cmake
   pool = console
 
 
 #############################################
 # A missing CMake input file is not an error.
 
-build C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeDetermineSystem.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeGenericSystem.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeInitializeConfigs.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystem.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystemSpecificInformation.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystemSpecificInitialize.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\RepositoryInfo.txt.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\cfgcmd.txt.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\gitclone.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\gitupdate.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\mkdirs.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\Platform\Windows.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\Platform\WindowsPaths.cmake CMakeCache.txt CMakeFiles\3.26.4\CMakeSystem.cmake CMakeLists.txt raylib_cpp-populate-prefix\tmp\raylib_cpp-populate-mkdirs.cmake: phony
+build C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeGenericSystem.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeInitializeConfigs.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystemSpecificInformation.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\CMakeSystemSpecificInitialize.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\RepositoryInfo.txt.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\cfgcmd.txt.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\gitclone.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\gitupdate.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\ExternalProject\mkdirs.cmake.in C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\Platform\Windows.cmake C$:\Program$ Files\JetBrains\CLion$ 2023.2.2\bin\cmake\win\x64\share\cmake-3.26\Modules\Platform\WindowsPaths.cmake CMakeCache.txt CMakeFiles\3.26.4\CMakeSystem.cmake CMakeLists.txt raylib_cpp-populate-prefix\tmp\raylib_cpp-populate-mkdirs.cmake: phony
 
 
 #############################################
