--- drawing.h
@@ -1,16 +1,0 @@
-//
-// Created by Craft on 11/4/2023.
-//
-
-#ifndef GITHUBGAMEOFF_DRAWING_H
-#define GITHUBGAMEOFF_DRAWING_H
-#pragma once
-#include "raylib-cpp.hpp"
-#include "gameobject.h"
-
-void DrawingExploreState(GameObject& player);
-void DrawingInStartState();
-void DrawingInEndState();
-void DrawingInMenuState();
-
-#endif //GITHUBGAMEOFF_DRAWING_H
