+++ pipe.h
@@ -1,0 +1,13 @@
+//
+// Created by Craft on 11/9/2023.
+//
+
+#ifndef GITHUBGAMEOFF_PIPE_H
+#define GITHUBGAMEOFF_PIPE_H
+#include "gameobject.h"
+
+struct Pipe: public GameObject{
+
+};
+
+#endif //GITHUBGAMEOFF_PIPE_H
